Account for different displays between first and subsequent logins.
On first login, the welcome banner is displayed when visiting the homepage.
On subsequents logins, however, a different banner is displayed, reminding
the user to select an update channel. We thus need to create a new needle
for that banner. We still use the 'welcome_banner' tag to simplify the
assert(-and-click) calls.
On first login, Firefox will throw a security alert due to our use of
self-signed certificates. Once this security exception is added, this
alert is not shown to the user anymore so we need to account for that.
Simply use a FIRST_BROWSE environment variable to distinguish first from
subsequent logins.