Improve pool scrub finished detection and import prior pool
Previously, the detection of a finished pool scrub was too relaxed
and could lead to a scrub listed as Running detected as Finished.
Include new needles that more specifically search for the finished
status.

Similarly, the detection of the tabs on the Pool page could occur
too quickly, thereby leading openQA to click on a non-clickable link
as we were already on the target page before the page reload.
Implement a wait_screen_change_and a brief sleep to ensure page
reload.

Add needles and correct unnecessary assert_and_click for pool delete.