Begin work on a configuration generator
To ensure the correctness of openQA's configuration with regards to
machines, test suites, and job groups/templates, we can use a scripted
generator as used in https://github.com/OSInside/kiwi-functional-tests

This commit includes the first attempt at doing so for a machine. First,
GET all machines from server, DELETE any existing machine with a matching
name, and then POST a new machine with our settings.