Add test_suite constant generator for First boot and Pools tests
Our tests can be organized in different domains, each with a set of
test suites (supportserver + webui), and with different levels.
Storage tests targeting Pools, for instance, can be viewed as the
repetition of the same combination of supportserver + webui for
each raid level supported.
This commit implements the beginning of a generator to simplify the
creation of such repeated test suites.