Add generation of job groups
Contrary to test suites and machines, job groups are rather complex
to write automatically from scratch. Let's thus take a different
approach and let the user define them by creating the files manually
to disk following a proper folder architecture. These files should
be placed in the "job_groups" directory, located at the root of this
repository, and with a max directory depth of 1 sub-directory. We
then read them from disk to establish:
  - parent group(s) if any
  - job groups
  - job groups templates