Correct missing relaxed matching parameters on pool balance needles
The "match" and "margin" settings for the needles with the
pool_detail_balance_running and pools_detail_balance_finished tags
were erroneously still set to the default values while they should
be relaxed a bit. Correct their JSON files to explicitly specify
more relaxed settings for match percentage as needed for the
observed variation in column headers on Pools page.