Use the config generator to establish products
Products/mediums were the only config that was not supported by the
configuration generator.

Define the set of products for each architecture we support. While
only x86_64 is supported at the moment, this may expand to arch64
in the future.
Establish the products by updating (PUT) or creating (POST) new
product(s) as required based on what is currently fetched from the
openQA server. Note that this process should happen before
establishing job groups as the latter validate the template(s) based
on the products defined in the database.